+++ pixels-common/src/main/java/io/pixelsdb/pixels/common/physical/Scheduler.java
@@ -17,9 +17,7 @@
  * License along with Pixels.  If not, see
  * <https://www.gnu.org/licenses/>.
  */
-package io.pixelsdb.pixels.common.physical.scheduler;
-
-import io.pixelsdb.pixels.common.physical.PhysicalReader;
+package io.pixelsdb.pixels.common.physical;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
