+++ pixels-common/src/main/java/io/pixelsdb/pixels/common/physical/io/PhysicalLocalReader.java
@@ -17,10 +17,11 @@
  * License along with Pixels.  If not, see
  * <https://www.gnu.org/licenses/>.
  */
-package io.pixelsdb.pixels.common.physical.impl;
+package io.pixelsdb.pixels.common.physical.io;
 
 import io.pixelsdb.pixels.common.physical.PhysicalReader;
 import io.pixelsdb.pixels.common.physical.Storage;
+import io.pixelsdb.pixels.common.physical.storage.LocalFS;
 
 import java.io.IOException;
 import java.io.RandomAccessFile;
