+++ pixels-common/src/main/java/io/pixelsdb/pixels/common/physical/scheduler/RateLimitedScheduler.java
@@ -1,7 +1,27 @@
+/*
+ * Copyright 2021 PixelsDB.
+ *
+ * This file is part of Pixels.
+ *
+ * Pixels is free software: you can redistribute it and/or modify
+ * it under the terms of the Affero GNU General Public License as
+ * published by the Free Software Foundation, either version 3 of
+ * the License, or (at your option) any later version.
+ *
+ * Pixels is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * Affero GNU General Public License for more details.
+ *
+ * You should have received a copy of the Affero GNU General Public
+ * License along with Pixels.  If not, see
+ * <https://www.gnu.org/licenses/>.
+ */
 package io.pixelsdb.pixels.common.physical.scheduler;
 
 import com.google.common.util.concurrent.RateLimiter;
 import io.pixelsdb.pixels.common.physical.PhysicalReader;
+import io.pixelsdb.pixels.common.physical.Scheduler;
 import io.pixelsdb.pixels.common.physical.Storage;
 import io.pixelsdb.pixels.common.utils.ConfigFactory;
 import org.apache.logging.log4j.LogManager;
@@ -23,13 +43,23 @@ import java.util.List;
 public class RateLimitedScheduler extends SortMergeScheduler
 {
     private static Logger logger = LogManager.getLogger(RateLimitedScheduler.class);
+    private static RateLimitedScheduler instance;
+
+    public static Scheduler Instance()
+    {
+        if (instance == null)
+        {
+            instance = new RateLimitedScheduler();
+        }
+        return instance;
+    }
 
     private RateLimiter mbpsRateLimiter;
     private RateLimiter rpsRateLimiter;
     private RetryPolicy retryPolicy;
     private final boolean enableRetry;
 
-    RateLimitedScheduler()
+    protected RateLimitedScheduler()
     {
         ConfigFactory.Instance().registerUpdateCallback("read.request.rate.limit.mbps", value ->
         {
